+++ NEWS.md
@@ -9,6 +9,7 @@ version 0.1.1
 =============
 
 * add more examples to non-internal functions and *do* run them (per CRAN)
+* added checks in the tests to ensure all exported functions have examples
 
 version 0.1.0
 =============